pyproject.toml
[project]
name = "mail_processor"
version = "0.1.0"
description = "Processing mails with rules"
authors = [
    {name = "Joshua", email = ""},
]
dependencies = [
    "google-api-python-client>=2.149.0",
    "pydantic>=2.9.2",
    "google-auth-oauthlib>=1.2.1",
    "pydantic-settings>=2.6.0",
    "rich>=13.9.3",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
